+++ jumpscale/packages/marketplace/chats/flist.py
@@ -1,7 +1,7 @@
 import math
 
 from jumpscale.loader import j
-from jumpscale.packages.tfgrid_solutions.chats.flist_deploy import FlistDeploy as BaseFlistDeploy
+from jumpscale.packages.tfgrid_solutions.chats.flist import FlistDeploy as BaseFlistDeploy
 from jumpscale.sals.chatflows.chatflows import chatflow_step
 from jumpscale.sals.marketplace import MarketPlaceChatflow, deployer, solutions
