+++ jumpscale/packages/marketplace/chats/gitea.py
@@ -1,4 +1,4 @@
-from jumpscale.packages.tfgrid_solutions.chats.gitea_deploy import GiteaDeploy as BaseGiteaDeploy
+from jumpscale.packages.tfgrid_solutions.chats.gitea import GiteaDeploy as BaseGiteaDeploy
 from jumpscale.sals.chatflows.chatflows import chatflow_step
 from jumpscale.sals.marketplace import MarketPlaceChatflow, deployer, solutions
