+++ jumpscale/packages/marketplace/chats/gollum.py
@@ -1,6 +1,6 @@
 import math
 
-from jumpscale.packages.tfgrid_solutions.chats.gollum_deploy import GollumDeploy as BaseGollumDeploy
+from jumpscale.packages.tfgrid_solutions.chats.gollum import GollumDeploy as BaseGollumDeploy
 from jumpscale.sals.chatflows.chatflows import chatflow_step
 from jumpscale.sals.marketplace import MarketPlaceChatflow, deployer, solutions
