+++ jumpscale/packages/marketplace/chats/minio.py
@@ -2,7 +2,7 @@ import math
 
 from jumpscale.clients.explorer.models import DiskType
 from jumpscale.loader import j
-from jumpscale.packages.tfgrid_solutions.chats.minio_deploy import MinioDeploy as BaseMinioDeploy
+from jumpscale.packages.tfgrid_solutions.chats.minio import MinioDeploy as BaseMinioDeploy
 from jumpscale.sals.chatflows.chatflows import chatflow_step
 from jumpscale.sals.marketplace import MarketPlaceChatflow, deployer, solutions
