+++ jumpscale/packages/marketplace/chats/network.py
@@ -1,5 +1,5 @@
 from jumpscale.loader import j
-from jumpscale.packages.tfgrid_solutions.chats.network_deploy import NetworkDeploy as BaseNetworkDeploy
+from jumpscale.packages.tfgrid_solutions.chats.network import NetworkDeploy as BaseNetworkDeploy
 from jumpscale.sals.chatflows.chatflows import StopChatFlow, chatflow_step
 from jumpscale.sals.marketplace import MarketPlaceChatflow, deployer, solutions
