+++ jumpscale/packages/marketplace/chats/ubuntu_simple.py
@@ -1,4 +1,3 @@
-from jumpscale.packages.tfgrid_solutions.chats.cryptpad_deploy import CryptpadDeploy as BaseCryptpadDeploy
 from jumpscale.sals.chatflows.chatflows import chatflow_step, StopChatFlow
 from jumpscale.sals.marketplace import MarketPlaceChatflow, deployer, solutions
 import uuid
